me/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pillow==11.3.0
psycopg2-binary==2.9.10
pyperclip==1.10.0
python-dotenv==1.1.1
screeninfo==0.8.1
SQLAlchemy==2.0.43
sswg==4.3.0
typing_extensions==4.15.0
ursina==8.1.1
Werkzeug==3.1.3
